requirements.txt
Flask==3.0.0
mysql-connector-python==8.2.0
pytest==7.4.3
pytest-flask==1.3.0
python-dotenv==1.0.0
